removed unused code, removed user_task model that has been replaced by answer, namespaced api controllers for best practises, updated tamplates, updated fields
