+++ app/controllers/api/v1/tasks_controller.rb
@@ -1,0 +1,17 @@
+class Api::V1::TasksController < Api::V1::BaseController
+  before_action :set_task, only: [:show]
+  
+  # GET /tasks/1
+  def show
+  end
+
+  private
+  
+    def set_task
+      @task = Task.find(params[:id])
+    end
+
+    def set_default_response_format
+      request.format = :json
+    end
+end
