+++ app/helpers/api/v1/base_helper.rb
@@ -1,0 +1,2 @@
+module Api::V1::BaseHelper
+end
