+++ app/models/answer.rb
@@ -1,0 +1,7 @@
+class Answer
+  include Mongoid::Document
+  field :message, type: String
+
+  belongs_to :user
+  belongs_to :task
+end
