+++ app/models/task.rb
@@ -2,14 +2,13 @@ class Task
   include Mongoid::Document
   include Mongoid::Timestamps
   
-  field :status, type: String
-  field :delivery_date, type: Time
   field :title, type: String
   field :description, type: String
   field :attachment, type: String
-  field :course_id, type: String
+  field :opened_date, type: DateTime
+  field :delivery_date, type: DateTime
 
   belongs_to :course
 
-  has_many :user_tasks
+  has_many :answers
 end
