+++ config/routes.rb
@@ -1,11 +1,16 @@
 Rails.application.routes.draw do
   mount RailsAdmin::Engine => '/admin', as: 'rails_admin'
   
-  resources :users, :only => [:create, :update] do 
-    member do
-      post :request_access_token
-      get :tasks
+  namespace :api do
+    namespace :v1 do
+      resources :users, :only => [] do 
+        member do
+          post :request_access_token
+          get :tasks
+        end
+      end
+      resources :tasks, :only => [:show]
     end
   end
-  resources :tasks, :only => [:index, :show]
+
 end
