+++ test/controllers/api/v1/base_controller_test.rb
@@ -1,6 +1,6 @@
 require 'test_helper'
 
-class UserTaskTest < ActiveSupport::TestCase
+class Api::V1::BaseControllerTest < ActionController::TestCase
   # test "the truth" do
   #   assert true
   # end
