+++ test/models/answer_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class AnswerTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
